WIP Improve variable names to distinguish decoded/serializer info
TODO: Better word commit message.

This change makes it clear whether the variables stores information
about the serialized data that was decoded, or about the serializable
type.

Looking back on the old encoder/decoder code, it was difficult following
the serialization/validation logic since it isn't immediately obvious
what the values were used for. (Whether it be a detail that is to be
serialized, or a value that was already decoded)