Change builder constructors to take a nullable `from`
There were design issues around providing an `NbtFormat` instance as the
default since not all the formats *can* have a default instance.
Previously this was worked around by having an illegal instance with
values that would be explicitly ignored by the builder using a
reference check.

The previous approach feels like a code smell, treating an incompletely
configured instance as if it were complete when passing it as a default,
while requiring special treatment for the "default" instance since
it isn't a behavioral subtype semantically.

With more specific `NbtFormat` subtypes coming later on, this new
approach with objects that provide the defaults should be more wieldy to
implement and maintain. The `Default` implementations that required
special handling have been removed, and there is now a single source of
truth for each option's default value.